Allow static methods on components
Having a static method (e.g. `fetchData`) to fetch just the data required to render a component on the server is a standard practice for universal apps.

`renderProps: Object` as currently declared doesn't allow the static method, because `Object` doesn't have them. Declaring it as any will fix this problem.